+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/ThreadSleepBlocking.kt
@@ -7,7 +7,7 @@ import org.pastalab.fray.rmi.ThreadState
 class ThreadSleepBlocking(val context: ThreadContext) : TimedBlockingOperation(true) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     context.pendingOperation = ThreadResumeOperation(type != InterruptionType.TIMEOUT)
-    context.state = ThreadState.Enabled
+    context.state = ThreadState.Runnable
     return null
   }
 }